+++ packages/edit-site/src/components/editor/index.js
@@ -22,7 +22,7 @@ import { store as noticesStore } from '@wordpress/notices';
 import { privateApis as routerPrivateApis } from '@wordpress/router';
 import { store as preferencesStore } from '@wordpress/preferences';
 import { decodeEntities } from '@wordpress/html-entities';
-import { Icon, directionSE } from '@wordpress/icons';
+import { Icon, directionNW } from '@wordpress/icons';
 
 /**
  * Internal dependencies
@@ -291,7 +291,7 @@ export default function EditSiteEditor( { isPostsList = false } ) {
 											) }
 											variants={ toggleHomeIconVariants }
 										>
-											<Icon icon={ directionSE } />
+											<Icon icon={ directionNW } />
 										</motion.div>
 									</motion.div>
 								)