+++ packages/icons/src/library/direction-se.js
@@ -5,7 +5,7 @@ import { SVG, Path } from '@wordpress/primitives';
 
 const directionSE = (
 	<SVG xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-		<Path d="M9.857 17.648v-1.5h5.26L5.792 6.822 6.852 5.76l9.326 9.326v-5.26h1.5v7.82h-7.82Z" />
+		<Path d="M10 18h8v-8h-1.5v5.5L7 6 6 7l9.5 9.5H10V18Z" />
 	</SVG>
 );