+++ packages/icons/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### New Features
 
 -   Add new `bell` and `bell-unread` icons.
--   Add new `directionSE` icon.
+-   Add new `directionSE` and `directionNW` icons.
 
 ## 10.7.0 (2024-09-05)